add pr description comparison
